openstack: Deprecate flag and use data_disk_size
The flag openstack_volume_flag (ovs) has been deprecated.  The standard
command line flag that should be used is data_disk_size (dds).

To allow the assignment of ovs value to dds, a disk spec object with
a specific 'disk_type' value is needed.  Used an abstract class
to overload the DiskSpec class in disk.py.

The cases identified and handled are:
* Case 1: ovs passed, dds not passed -> dds gets ovs value
* Case 2: ovs not passed, dds passed -> dds value used
* Case 3: ovs passed, dds passed -> dds value used